refactor: move time tracking components into features/timeTracking
Flatten TimeEntries/** and TimeTrackingStats into one directory per
component under features/timeTracking, per the components SKILL layout.
