+++ src/components/features/timeTracking/ActiveTimeTracker/ActiveTimeTracker.tsx
@@ -9,8 +9,8 @@ import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { VStack } from '@ui/Stack/Stack'
 import { Container } from '@components/Container/Container'
-import { ActiveTimeTrackerBanner } from '@components/TimeEntries/ActiveTimeTracker/ActiveTimeTrackerBanner'
-import { ActiveTimeTrackerStartDrawer } from '@components/TimeEntries/ActiveTimeTracker/ActiveTimeTrackerStartDrawer'
+import { ActiveTimeTrackerBanner } from '@features/timeTracking/ActiveTimeTracker/ActiveTimeTrackerBanner'
+import { ActiveTimeTrackerStartDrawer } from '@features/timeTracking/ActiveTimeTracker/ActiveTimeTrackerStartDrawer'
 
 import './activeTimeTracker.scss'
