+++ src/components/features/timeTracking/ActiveTimeTracker/activeTimeTracker.scss
