+++ src/components/features/timeTracking/ServiceArchiveModal/ServiceArchiveModal.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { type CatalogService } from '@schemas/catalogService'
 import { usePostArchiveCatalogService } from '@api/businesses/[business-id]/catalog/services/[service-id]/archive/post'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
-import { BaseConfirmationModal } from '@components/blocks/BaseConfirmationModal/BaseConfirmationModal'
+import { BaseConfirmationModal } from '@blocks/BaseConfirmationModal/BaseConfirmationModal'
 
 type ServiceArchiveModalProps = {
   service: CatalogService | null
