+++ src/components/features/timeTracking/TimeEntryForm/TimeEntryForm.tsx
@@ -13,8 +13,8 @@ import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Form } from '@ui/Form/Form'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { CustomerSelector } from '@components/CustomerSelector/CustomerSelector'
-import { useTimeEntryForm } from '@components/TimeEntries/TimeEntryForm/useTimeEntryForm'
-import { TimeEntryServiceSelector } from '@components/TimeEntries/TimeEntryServiceSelector/TimeEntryServiceSelector'
+import { useTimeEntryForm } from '@features/timeTracking/TimeEntryForm/useTimeEntryForm'
+import { TimeEntryServiceSelector } from '@features/timeTracking/TimeEntryServiceSelector/TimeEntryServiceSelector'
 
 import './timeEntryForm.scss'
