+++ src/components/features/timeTracking/TimeTrackingServicesDrawer/TimeTrackingServicesDrawer.tsx
@@ -16,12 +16,11 @@ import { Span } from '@ui/Typography/Text'
 import { ExpandableCard } from '@blocks/ExpandableCard/ExpandableCard'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
 import { ConditionalList } from '@components/utility/ConditionalList'
+import { ServiceArchiveModal } from '@features/timeTracking/ServiceArchiveModal/ServiceArchiveModal'
+import { ServiceFormCard } from '@features/timeTracking/ServiceFormCard/ServiceFormCard'
+import { ServiceRestoreModal } from '@features/timeTracking/ServiceRestoreModal/ServiceRestoreModal'
 
 import './timeTrackingServicesDrawer.scss'
-
-import { ServiceArchiveModal } from './ServiceArchiveModal'
-import { ServiceFormCard } from './ServiceFormCard'
-import { ServiceRestoreModal } from './ServiceRestoreModal'
 
 type ServicesTab = 'active' | 'archived'
