+++ src/views/TimeTracking.tsx
@@ -7,10 +7,10 @@ import { useGlobalDateRange } from '@providers/DateStoreProvider/GlobalDateStore
 import { TimeTrackingServicesDrawerProvider, useTimeTrackingServicesDrawer } from '@providers/TimeTrackingServicesDrawerProvider/TimeTrackingServicesDrawerProvider'
 import { type DropdownMenuItem } from '@ui/DropdownMenu/DropdownMenu'
 import { DataTableHeaderMenu } from '@blocks/DataTable/DataTableHeaderMenu'
-import { ActiveTimeTracker } from '@components/TimeEntries/ActiveTimeTracker/ActiveTimeTracker'
-import { TimeEntries } from '@components/TimeEntries/TimeEntries'
-import { TimeTrackingStats } from '@components/TimeTrackingStats/TimeTrackingStats'
 import { View } from '@components/View/View'
+import { ActiveTimeTracker } from '@features/timeTracking/ActiveTimeTracker/ActiveTimeTracker'
+import { TimeEntries } from '@features/timeTracking/TimeEntries/TimeEntries'
+import { TimeTrackingStats } from '@features/timeTracking/TimeTrackingStats/TimeTrackingStats'
 
 export interface TimeTrackingStringOverrides {
   title?: string
